Minor improvements and changes
The UserForm fMsg  cannot  be used directly but via the Box and Dsply services provided by the mMsg module. The Dsply service now allows to display the message modeless. If modeless the Dsply service waits until a button is clicked and passes the result on to the caller.
The Deomos, the Tests and the Regression Test has been reworked.