Total flexibility not implemented (activated respectively)
The approach looked very promising until it failed to Add a Frame control to a Frame. Excel crashed and there is no solution for that. I simply had to give up with the idea to create all controls dynamically. The code remains prepared to to so however. Once a solution for the problem becomes available all design can be removed.